Adjust to Gson @Expose and move sweetness and toppings
