+++ ooc-spring/src/main/java/com/mapringg/bab/models/Customer.java
@@ -6,7 +6,7 @@ import javax.persistence.*;
  * @author mapring
  */
 @Entity
-@Inheritance
+@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
 public class Customer {
 
     @Id
