+++ ooc-spring/src/main/java/com/mapringg/bab/models/MenuType.java
@@ -6,12 +6,15 @@ import javax.persistence.*;
  * @author mapring
  */
 @Entity
-@Inheritance
+@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
 public class MenuType {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column(insertable = false, updatable = false)
+    private String dtype;
 
     public Long getId() {
         return id;
@@ -19,5 +22,13 @@ public class MenuType {
 
     public void setId(Long id) {
         this.id = id;
+    }
+
+    public String getDtype() {
+        return dtype;
+    }
+
+    public void setDtype(String dtype) {
+        this.dtype = dtype;
     }
 }
